comit 10.01.2025
gestion compte bancaire
gestion transaction
soucis sur coté virement

+++ database/migrations/2024_12_20_091713_create_compte_bancaire_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('numero_compte')->unique();
             $table->unsignedInteger('idUser');
             $table->string('iban');
-            $table->integer('budget')->default(0);
+            $table->decimal('budget', 10, 2) ->default(0);
             $table->string('statut')->default('ouvert');
             $table->string('typeCompte')->default('Courant');
             $table->timestamps();
